# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: SafeSchool safeguarding and incident reporting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /protocols:
    get:
      operationId: listProtocols
      tags: [protocols]
      summary: List protocols
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated protocols
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProtocols"
    post:
      operationId: createProtocol
      tags: [protocols]
      summary: Open a new protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProtocolBody"
      responses:
        "201":
          description: Protocol created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
  /protocols/{id}:
    get:
      operationId: getProtocol
      tags: [protocols]
      summary: Get protocol detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Protocol detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProtocol
      tags: [protocols]
      summary: Update protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProtocolBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocol"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Incident:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        schoolId:
          type: string
        reporterId:
          type: ["string", "null"]
        reporterRole:
          type: string
        anonymous:
          type: boolean
        category:
          type: string
        escalationTier:
          type: integer
        safeguardingTrigger:
          type: boolean
        incidentDate:
          type: string
        incidentTime:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        victimIds:
          type: array
          items:
            type: string
        perpetratorIds:
          type: array
          items:
            type: string
        personInvolvedText:
          type: ["string", "null"]
        witnessIds:
          type: array
          items:
            type: string
        witnessText:
          type: ["string", "null"]
        emotionalState:
          type: ["string", "null"]
        emotionalFreetext:
          type: ["string", "null"]
        happeningToMe:
          type: boolean
        happeningToSomeoneElse:
          type: boolean
        iSawIt:
          type: boolean
        childrenSeparated:
          type: ["boolean", "null"]
        coordinatorNotified:
          type: ["boolean", "null"]
        immediateActionTaken:
          type: ["string", "null"]
        partOfKnownPattern:
          type: ["boolean", "null"]
        toldByChild:
          type: ["boolean", "null"]
        childConsentToShare:
          type: ["boolean", "null"]
        formalResponseRequested:
          type: boolean
        requestExternalReferral:
          type: boolean
        confidentialFlag:
          type: boolean
        status:
          type: string
        protocolId:
          type: ["string", "null"]
        createdAt:
          type: string
        reporterName:
          type: ["string", "null"]
        victimNames:
          type: array
          items:
            type: string
        perpetratorNames:
          type: array
          items:
            type: string
      required:
        - id
        - referenceNumber
        - schoolId
        - reporterRole
        - anonymous
        - category
        - escalationTier
        - safeguardingTrigger
        - incidentDate
        - status
        - createdAt
...
    CreateProtocolBody:
      type: object
      properties:
        protocolType:
          type: string
        protocolSource:
          type: ["string", "null"]
        genderBasedViolence:
          type: boolean
          default: false
        context:
          type: ["string", "null"]
        linkedIncidentIds:
          type: array
          items:
            type: string
        victimId:
          type: string
        allegedPerpetratorIds:
          type: array
          items:
            type: string
        riskAssessment:
          type: ["string", "null"]
        protectiveMeasures:
          type: array
          items:
            type: string
        externalReferralRequired:
          type: boolean
          default: false
        externalReferralBody:
          type: ["string", "null"]
      required:
        - protocolType
        - victimId
    Protocol:
      type: object
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        schoolId:
          type: string
        openedBy:
          type: string
        openedAt:
          type: string
        protocolType:
          type: string
        protocolSource:
          type: ["string", "null"]
        genderBasedViolence:
          type: boolean
        context:
          type: ["string", "null"]
        linkedIncidentIds:
          type: array
          items:
            type: string
        victimId:
          type: string
        allegedPerpetratorIds:
          type: array
          items:
            type: string
        parentNotificationSent:
          type: boolean
        interviewsRequired:
          type: boolean
        riskAssessment:
          type: ["string", "null"]
        protectiveMeasures:
          type: array
          items:
            type: string
        externalReferralRequired:
          type: boolean
        externalReferralBody:
          type: ["string", "null"]
        status:
          type: string
        resolutionNotes:
          type: ["string", "null"]
        closedAt:
          type: ["string", "null"]
        updatedAt:
          type: string
        victimName:
          type: ["string", "null"]
        openedByName:
          type: ["string", "null"]
      required:
        - id
        - referenceNumber
        - schoolId
        - openedBy
        - openedAt
        - protocolType
        - genderBasedViolence
        - victimId
        - parentNotificationSent
        - interviewsRequired
        - externalReferralRequired
        - status
        - updatedAt
    ProtocolDetail:
      type: object
      properties:
        protocol:
          $ref: "#/components/schemas/Protocol"
        linkedIncidents:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        interviews:
          type: array
          items:
            $ref: "#/components/schemas/Interview"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
      required:
        - protocol
        - linkedIncidents
        - interviews
        - notifications
    Interview:
      type: object
      properties:
        id:
          type: string
        protocolId:
          type: string
        intervieweeId:
          type: string
        intervieweeRole:
          type: string
        conductedBy:
          type: string
        interviewDate:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        annexReference:
          type: ["string", "null"]
        createdAt:
          type: string
        intervieweeName:
          type: ["string", "null"]
        conductedByName:
          type: ["string", "null"]
      required:
        - id
        - protocolId
        - intervieweeId
        - intervieweeRole
        - conductedBy
        - createdAt
    PaginatedProtocols:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Protocol"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    UpdateProtocolBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        riskAssessment:
          type: ["string", "null"]
        protectiveMeasures:
          type: array
          items:
            type: string
        resolutionNotes:
          type: ["string", "null"]
        parentNotificationSent:
          type: ["boolean", "null"]
        externalReferralRequired:
          type: ["boolean", "null"]
        externalReferralBody:
          type: ["string", "null"]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        recipientId:
          type: string
        trigger:
          type: string
        channel:
          type: string
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        sentAt:
          type: string
        acknowledgedAt:
          type: ["string", "null"]
        delivered:
          type: boolean
      required:
        - id
        - schoolId
        - recipientId
        - trigger
        - channel
        - sentAt
        - delivered
...